pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"



[project]
name = "solara"
authors = [{name = "Maarten A. Breddels", email = "maartenbreddels@gmail.com"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
packages = [
    { include = "solara" }
]
dependencies = [
    "reacton>=0.12.0",
    "ipywidgets<8",
    "diskcache",
    "markdown",
    "pygments",
    "pymdown-extensions",
    "markdown-it-py",
    "mdit-py-plugins",
    "humanize",
    "ipyvuetify",
    "ipyvue>=1.8.0",
    "numpy",
    "pillow",
]

[tool.flit.external-data]
directory = "prefix"

[tool.flit.module]
name = "solara"



[project.optional-dependencies]
dev = [
    "flake8",
    "bqplot",
    "bqplot-image-gl",
    "vaex-core",
    "vaex-hdf5",
    "vaex-jupyter",
    "black",
    "mypy",
    "playwright",
    "plotly",
    "pytest-playwright",
    "types-requests",
    "types-markdown",
    "types-PyYAML",
    "pytest",
    "pytest-cov",
    "pytest-timeout",
    "pre-commit",
    "bump2version",
    "watchdog",
]
assets = [
    "solara-assets==0.10.4"
]
server = [
    "jinja2",
    "pydantic",
    "click",
    "rich_click",
    "uvicorn",
    "watchgod",
    "websockets",
    "starlette",
    "nbconvert",
    "jupyter_server",
    "jupyter_client>=7.0.0",
]
flask = [
    "flask",
    "flask-sock",
]
examples = [
    "numpy",
    "bqplot",
    "altair",
    "ipycanvas",
    "vega_datasets",
    "plotly",
    "vaex-core",
    "vaex-hdf5",
]

[project.urls]
Home = "https://www.github.com/widgetti/solara"

[project.scripts]
solara = "solara.__main__:main"

[tool.black]
line-length = 160

[tool.isort]
profile = "black"


[tool.pytest.ini_options]
addopts = "--disable-warnings --screenshot=only-on-failure"
timeout = 180
python_files = "*.py"
